+++ activiti-cloud-query-service/activiti-cloud-starter-query/src/test/java/org/activiti/cloud/starter/tests/QueryAdminProcessDefinitionIT.java
@@ -72,7 +72,7 @@ public class QueryAdminProcessDefinitionIT {
 
     @BeforeEach
     public void setUp() {
-        identityTokenProducer.setTestUser("hradmin");
+        identityTokenProducer.withTestUser("hradmin");
     }
 
     @AfterEach
