+++ activiti-cloud-query-service/activiti-cloud-starter-query/src/test/java/org/activiti/cloud/starter/tests/QueryApplicationEntityIT.java
@@ -81,7 +81,7 @@ public class QueryApplicationEntityIT {
 
     @Test
     public void shouldGetDeployedApplicationsWhenUserIsAdmin() {
-        identityTokenProducer.setTestUser("hradmin");
+        identityTokenProducer.withTestUser("hradmin");
         getDeployedApplication(ADMIN_APPS_URL);
     }
