+++ activiti-cloud-query-service/activiti-cloud-starter-query/src/test/java/org/activiti/cloud/starter/tests/QueryBPMNActivityIT.java
@@ -89,7 +89,7 @@ public class QueryBPMNActivityIT {
 
     @BeforeEach
     public void setUp() throws IOException {
-        identityTokenProducer.setTestUser("hruser");
+        identityTokenProducer.withTestUser("hruser");
 
         eventsAggregator = new EventsAggregator(producer);
