+++ activiti-cloud-query-service/activiti-cloud-starter-query/src/test/java/org/activiti/cloud/starter/tests/QueryProcessDiagramIT.java
@@ -102,7 +102,7 @@ public class QueryProcessDiagramIT {
 
     @BeforeEach
     public void setUp() throws IOException {
-        identityTokenProducer.setTestUser("hruser");
+        identityTokenProducer.withTestUser("hruser");
 
         eventsAggregator = new EventsAggregator(producer);
