+++ activiti-cloud-runtime-bundle-service/activiti-cloud-starter-runtime-bundle/src/test/java/org/activiti/cloud/starter/tests/runtime/AbstractMQServiceTaskIT.java
@@ -86,7 +86,7 @@ public abstract class AbstractMQServiceTaskIT {
 
     @BeforeEach
     public void setUp() {
-        identityTokenProducer.setTestUser(keycloakTestUser);
+        identityTokenProducer.withTestUser(keycloakTestUser);
     }
 
     @Test
