+++ activiti-cloud-runtime-bundle-service/activiti-cloud-starter-runtime-bundle/src/test/java/org/activiti/cloud/starter/tests/runtime/TaskVariableMappingIT.java
@@ -72,7 +72,7 @@ public class TaskVariableMappingIT {
 
     @BeforeEach
     public void setUp() {
-        identityTokenProducer.setTestUser(keycloakTestUser);
+        identityTokenProducer.withTestUser(keycloakTestUser);
     }
 
     @Test
