+++ activiti-cloud-service-common/activiti-cloud-services-test/src/main/java/org/activiti/cloud/services/test/TestConfiguration.java
@@ -21,9 +21,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
-import org.activiti.cloud.identity.IdentityTokenProducer;
-import org.activiti.cloud.identity.interceptor.IdentityTokenInterceptor;
-import org.activiti.cloud.identity.keycloak.KeycloakTokenProducer;
+import org.activiti.cloud.services.test.identity.IdentityTokenProducer;
+import org.activiti.cloud.services.test.identity.interceptor.IdentityTokenInterceptor;
+import org.activiti.cloud.services.test.identity.keycloak.KeycloakTokenProducer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
